Many mobile phones have display defects.
The cross boxes in AndroidAPS, especially in the CONFIG are located at the border of the display. This area is quite sensitive for delaminations of the display. I have experienced that and had asked many questions why I coulden't select the cross box although it is selectable. at the end the root cause was the display.