Manual export settings
Base was https://www.facebook.com/groups/AndroidAPSUsers/permalink/2265623286992409/?comment_id=2266001030287968&reply_comment_id=2266361260251945&comment_tracking=%7B%22tn%22%3A%22R%2310%22%7D (see also proposed changes on Dana RS page).

Put export settings to the bottom of page as main topic is objectives. Is it possible to define an anchor and link for the top of the page to the bottom?

Not the best place (will ad also on Dana RS pump page) but on objectives page it is the most detailed settings export/import description I have found in wiki. Discussing with An Di if a separate page "Switch phone" would be helpful.