Separate page What you need to get started
Just accidently saw that this information can be found on the index page.
I think a separate page in the navigation menu on the left would be helpful for starters.
Changed layout to bullet list and added links to relevant Wiki pages.
Not sure if note is displayed correctly (blue background and i sign). Please check.