Update Open-APS-features.md
In "(...) hourly basal rate programmed in a user’s pump, or, if enabled, determined by autotune"  I deleted the last bit: "or, if enabled, determined by autotune". In sections for Advanced settings for SMB that describe the "Max daily safety multiplier" and "Current Basal safety multiplier".

In AndroidAPS, autotune will not automatically adjust the basal rate programmed in a user's pump (as opposed to OpenAPS) and therefore I consider that part redundant and confusing.