Update BG-Source.md
Addining aditional information about how to start sensor for G6. More detail information on how to configure xdrip for integration with AAPS and NS to provide following features from AAPS to xdrip: treatments on the main  screan (carbs, ecarbs, smb, bolus), TBR displayed in the upper part of the main screan, IOB and BGI displayed as the extra line in the uper left corner of the main screan.