Update BG-Source.md
Adding additional information about how to start sensor for G6. More detail information on how to configure xdrip for integration with AAPS and NS to provide following features from AAPS to xdrip: treatments on the main screen (carbs, ecarbs, smb, bolus), TBR displayed in the upper part of the main screen, IOB and BGI displayed as the extra line in the upper left corner of the main screen. Eliminating situation when duplicated treatments are stored in NS.