Dev branch separate page
As proposed by Adrian. Against the background of recent reports (i.e. DJ RA Ebert), I have emphasized the part that this branch is intended only for development and not for live operation. For the same reason I removed the direct link to the dev branch. Developers - and only these and testers are to be addressed - will surely find them without a link and it's harder to find for the DAO (stupidest assumable user).
On the background of latest reports (i.e. DJ RA Ebert) I put a bit more