TranslateAgain
PDF screenshot for wiki translators
Unfortunately there is a problem with Crowdin which we could not solve yet. When we change something in a page or add just a little information some of the translations already done are "lost" in Crowdin. Luckily they are not completely gone but need to be reactivated.