Update Children.rst
With the questions that are part of the future versions of AdroidAPS there is a bit of a confusion in the topic: monitoring children: how can you monitor AAPS of your child remotely? 
only NSClient can monitor AAPS; 
Xdrip+, Spike, Dexcom Follow are monitoring  BG's and giving  alarms but it would be wrong to regard them as remotely monitoring  AAPS.
So I have partially edited this page to introduce and explain NSClient app. 
I have also explained the other apps as tools to monitor BG and not AAPS ( since they do not monitor AAPS)
I have also include that Dexcom Follow can also be used from xDrip+ as well as from  the app Dexcom but with a note that there are other preferred options. 
Please consider the new headings:  methodology;  other tools and apps for partial monitoring 
Lastly the answers of the objective would need to be changed to only NSClient being correct, o rthe question rephrased to something like : "What tools are available to assist in the remote monitor of your child using AAPS?" and the answer would be all but "Loop app on iPhone ".