Update Accessing-logfiles.md
Thanks for your improvement. 
For standard dotted lists such as below the screenshot the empty line between list items is not necessary.
For a sub-list (as I did with your folder info) there must be one single line after the main dot and the sub-list must start with three spaces.