testing with absolute references instead of relative
relative references are not working with the RST file is imported into another file.  Sphinx says at runtime the text is blindly inserted and relative reference paths are not updated.  If so, then relative referenced image paths that function here will not function when this file is imported and referenced elsewhere.