Update FAQ.md
I noticed an error in the explanation of ISF in the FAQ for Loopers part of the docs:
It was explaining that a higher ISF leads to a smaller drop in BG for one unit of insulin and vice versa. I have edited the text to be more clear and made a distinction between wat the ISF means and what it leads to.