Update FAQ.md
added more context so it explains that a lower ISF is not more aggressive in and of itself but the loop will be because of it. If you drop from 200 to 100 the drop is 100. the amount of insulin needed for that is irrelevant. 
What is relevant is that a lower ISF means the loop will give more insulin