Update MedtronicPump.md
I rewrote the Medtronic documentation as part of an exercise to improve my github skills.  This is an attempt to clear up some of the language and try to help people setting up Medtronic pumps.  I thought it might be of wider use to others.  I've removed some of the language related to the initial release.