Update index.rst
As requested by @peterleimbach, some improvements on syntax for english documentation. My apologies, I believe that may be a link or two in the component/configurations that were not connecting to the expected places, but as I am not familiar with fixing links; I thought it best to leave them be for now.