Update introduction.md
Peter, I have tidied up the references in the clinician section. Please could you correct the first row of the table - Milos mentioned the fact that we have iAPS as being Android not iPhone in the large table of compatibility. I can't work out how to edit your table in Markdown, we just need a "x" in the android box and a "tick" in the iphone box. Thanks.