Update preparing.md
I have made some edits, but I think the difficulty is that this appears to be an index, and many of the sections haven't been written or moved to the master yet, so we don't really have the links. I think it is acceptable for now, we will need to revise it at the end.