Update remote-control.md
I have modified the heading priority, so that they look more ordered in the side panel (NSClient now the same level as SMS commands etc) in preparation for transfer to the master.
I have also added a link to the master smartwatch section. I did navigate into the master branch in the files to find the correct link, but the root of the name was unchanged from the rework project, so I hope this is correct, let me know if not.