Update following-only.md
This is a basic "following only" section. In time I would like to do a poll on FB and ask which methods are most commonly used for following, and then add in more detail about them in due course. I'd appreciate your guidance on if the collapsible section works OK. Thanks.