add terminal method
Add CL method for transfering wear-apk to smartwatch. 

Ive noticed that Easy fire tool and Wear installer can be a bit cumbersome sometimes. Although it requires the cli, this seems to be a reliable easy method that takes 5min. 

I havent tested this on windows as I dont have access to a windows machine.